# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CouchPotato - Personal movie and TV show tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EntryType = typeof EntryType[keyof typeof EntryType];


export const EntryType = {
  movie: 'movie',
  show: 'show',
} as const;

export type EntryStatus = typeof EntryStatus[keyof typeof EntryStatus];


export const EntryStatus = {
  watching: 'watching',
  plan_to_watch: 'plan_to_watch',
  completed: 'completed',
} as const;

export interface Entry {
  id: number;
  title: string;
  type: EntryType;
  status: EntryStatus;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  dateWatched?: string | null;
  /** @nullable */
  year: number | null;
  /**
     * @minimum 1
     * @maximum 5
     * @nullable
     */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  tmdbId?: number | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export type EntryInputType = typeof EntryInputType[keyof typeof EntryInputType];


export const EntryInputType = {
  movie: 'movie',
  show: 'show',
} as const;

export type EntryInputStatus = typeof EntryInputStatus[keyof typeof EntryInputStatus];


export const EntryInputStatus = {
  watching: 'watching',
  plan_to_watch: 'plan_to_watch',
  completed: 'completed',
} as const;

export interface EntryInput {
  /** @minLength 1 */
  title: string;
  type: EntryInputType;
  status?: EntryInputStatus;
  posterUrl?: string;
  dateWatched?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
  notes?: string;
  synopsis?: string;
  tmdbId?: number;
  tags?: string[];
}

export type EntryUpdateType = typeof EntryUpdateType[keyof typeof EntryUpdateType];


export const EntryUpdateType = {
  movie: 'movie',
  show: 'show',
} as const;

export type EntryUpdateStatus = typeof EntryUpdateStatus[keyof typeof EntryUpdateStatus];


export const EntryUpdateStatus = {
  watching: 'watching',
  plan_to_watch: 'plan_to_watch',
  completed: 'completed',
} as const;

export interface EntryUpdate {
  /** @minLength 1 */
  title?: string;
  type?: EntryUpdateType;
  status?: EntryUpdateStatus;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  dateWatched?: string | null;
  /**
     * @minimum 1
     * @maximum 5
     * @nullable
     */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  tmdbId?: number | null;
  tags?: string[];
}

export type TmdbResultType = typeof TmdbResultType[keyof typeof TmdbResultType];


export const TmdbResultType = {
  movie: 'movie',
  show: 'show',
} as const;

export interface TmdbResult {
  tmdbId: number;
  title: string;
  type: TmdbResultType;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  overview?: string | null;
}

export interface TmdbSearchResponse {
  results: TmdbResult[];
}

export interface TmdbPopularResponse {
  movies: TmdbResult[];
  shows: TmdbResult[];
}

export interface MonthCount {
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  count: number;
}

export interface TagCount {
  tag: string;
  count: number;
}

export interface YearStats {
  year: number;
  total: number;
  movies: number;
  shows: number;
  /** @nullable */
  averageRating: number | null;
  byMonth: MonthCount[];
  byTag: TagCount[];
}

export interface YearSummary {
  /** @nullable */
  year: number | null;
  count: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListEntriesParams = {
/**
 * Filter by year watched
 */
year?: number;
/**
 * Filter by type
 */
type?: ListEntriesType;
/**
 * Search by title
 */
search?: string;
/**
 * Minimum rating filter
 * @minimum 1
 * @maximum 5
 */
minRating?: number;
/**
 * Filter by watch status
 */
status?: ListEntriesStatus;
};

export type ListEntriesType = typeof ListEntriesType[keyof typeof ListEntriesType];


export const ListEntriesType = {
  movie: 'movie',
  show: 'show',
} as const;

export type ListEntriesStatus = typeof ListEntriesStatus[keyof typeof ListEntriesStatus];


export const ListEntriesStatus = {
  watching: 'watching',
  plan_to_watch: 'plan_to_watch',
  completed: 'completed',
} as const;

export type GetStatsParams = {
/**
 * Year to compute stats for
 */
year: number;
};

export type TmdbSearchParams = {
/**
 * Search query
 */
q: string;
};