# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CouchPotato - Personal movie and TV show tracker
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: entries
    description: Movie and TV show log entries
  - name: stats
    description: Viewing statistics
  - name: tmdb
    description: TMDB proxy for search and discovery

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List all log entries
      parameters:
        - name: year
          in: query
          schema:
            type: integer
          description: Filter by year watched
        - name: type
          in: query
          schema:
            type: string
            enum: [movie, show]
          description: Filter by type
        - name: search
          in: query
          schema:
            type: string
          description: Search by title
        - name: minRating
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
          description: Minimum rating filter
        - name: status
          in: query
          schema:
            type: string
            enum: [watching, plan_to_watch, completed]
          description: Filter by watch status
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Log a new movie or show
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryInput"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /entries/{id}:
    get:
      operationId: getEntry
      tags: [entries]
      summary: Get a single entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEntry
      tags: [entries]
      summary: Update an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryUpdate"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete an entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Entry deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get viewing stats for a year
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
          description: Year to compute stats for
      responses:
        "200":
          description: Year statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YearStats"

  /years:
    get:
      operationId: listYears
      tags: [entries]
      summary: List all years that have entries
      responses:
        "200":
          description: List of years with entry counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/YearSummary"

  /tmdb/search:
    get:
      operationId: tmdbSearch
      tags: [tmdb]
      summary: Search TMDB for movies and TV shows
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmdbSearchResponse"
        "503":
          description: TMDB API key not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tmdb/trending:
    get:
      operationId: tmdbTrending
      tags: [tmdb]
      summary: Get trending movies and shows from TMDB
      responses:
        "200":
          description: Trending results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmdbSearchResponse"
        "503":
          description: TMDB API key not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tmdb/popular:
    get:
      operationId: tmdbPopular
      tags: [tmdb]
      summary: Get popular movies and shows from TMDB
      responses:
        "200":
          description: Popular results split by type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TmdbPopularResponse"
        "503":
          description: TMDB API key not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Entry:
      type: object
      required:
        - id
        - title
        - type
        - status
        - year
        - tags
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [movie, show]
        status:
          type: string
          enum: [watching, plan_to_watch, completed]
          default: completed
        posterUrl:
          type: ["string", "null"]
        dateWatched:
          type: ["string", "null"]
          format: date
        year:
          type: ["integer", "null"]
        rating:
          type: ["integer", "null"]
          minimum: 1
          maximum: 5
        notes:
          type: ["string", "null"]
        synopsis:
          type: ["string", "null"]
        tmdbId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    EntryInput:
      type: object
      required:
        - title
        - type
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [movie, show]
        status:
          type: string
          enum: [watching, plan_to_watch, completed]
          default: completed
        posterUrl:
          type: string
        dateWatched:
          type: string
          format: date
        rating:
          type: integer
          minimum: 1
          maximum: 5
        notes:
          type: string
        synopsis:
          type: string
        tmdbId:
          type: integer
        tags:
          type: array
          items:
            type: string

    EntryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [movie, show]
        status:
          type: string
          enum: [watching, plan_to_watch, completed]
        posterUrl:
          type: ["string", "null"]
        dateWatched:
          type: ["string", "null"]
          format: date
        rating:
          type: ["integer", "null"]
          minimum: 1
          maximum: 5
        notes:
          type: ["string", "null"]
        synopsis:
          type: ["string", "null"]
        tmdbId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string

    TmdbResult:
      type: object
      required:
        - tmdbId
        - title
        - type
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [movie, show]
        year:
          type: ["integer", "null"]
        posterUrl:
          type: ["string", "null"]
        overview:
          type: ["string", "null"]

    TmdbSearchResponse:
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/TmdbResult"

    TmdbPopularResponse:
      type: object
      required:
        - movies
        - shows
      properties:
        movies:
          type: array
          items:
            $ref: "#/components/schemas/TmdbResult"
        shows:
          type: array
          items:
            $ref: "#/components/schemas/TmdbResult"

    YearStats:
      type: object
      required:
        - year
        - total
        - movies
        - shows
        - averageRating
        - byMonth
        - byTag
      properties:
        year:
          type: integer
        total:
          type: integer
        movies:
          type: integer
        shows:
          type: integer
        averageRating:
          type: ["number", "null"]
        byMonth:
          type: array
          items:
            $ref: "#/components/schemas/MonthCount"
        byTag:
          type: array
          items:
            $ref: "#/components/schemas/TagCount"

    MonthCount:
      type: object
      required:
        - month
        - count
      properties:
        month:
          type: integer
          minimum: 1
          maximum: 12
        count:
          type: integer

    TagCount:
      type: object
      required:
        - tag
        - count
      properties:
        tag:
          type: string
        count:
          type: integer

    YearSummary:
      type: object
      required:
        - year
        - count
      properties:
        year:
          type: ["integer", "null"]
        count:
          type: integer

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string