# lib/api-zod/src/generated/types/entry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CouchPotato - Personal movie and TV show tracker
 * OpenAPI spec version: 0.1.0
 */
import type { EntryStatus } from './entryStatus';
import type { EntryType } from './entryType';

export interface Entry {
  id: number;
  title: string;
  type: EntryType;
  status: EntryStatus;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  dateWatched?: Date | null;
  /** @nullable */
  year: number | null;
  /**
     * @minimum 1
     * @maximum 5
     * @nullable
     */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  tmdbId?: number | null;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}