# lib/api-zod/src/generated/types/entryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CouchPotato - Personal movie and TV show tracker
 * OpenAPI spec version: 0.1.0
 */
import type { EntryInputStatus } from './entryInputStatus';
import type { EntryInputType } from './entryInputType';

export interface EntryInput {
  /** @minLength 1 */
  title: string;
  type: EntryInputType;
  status?: EntryInputStatus;
  posterUrl?: string;
  dateWatched?: Date;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating?: number;
  notes?: string;
  synopsis?: string;
  tmdbId?: number;
  tags?: string[];
}