# lib/api-zod/src/generated/types/entryUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CouchPotato - Personal movie and TV show tracker
 * OpenAPI spec version: 0.1.0
 */
import type { EntryUpdateStatus } from './entryUpdateStatus';
import type { EntryUpdateType } from './entryUpdateType';

export interface EntryUpdate {
  /** @minLength 1 */
  title?: string;
  type?: EntryUpdateType;
  status?: EntryUpdateStatus;
  /** @nullable */
  posterUrl?: string | null;
  /** @nullable */
  dateWatched?: Date | null;
  /**
     * @minimum 1
     * @maximum 5
     * @nullable
     */
  rating?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  tmdbId?: number | null;
  tags?: string[];
}