# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CouchPotato - Personal movie and TV show tracker
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}