# lib/api-zod/src/generated/types/monthCount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * CouchPotato - Personal movie and TV show tracker
 * OpenAPI spec version: 0.1.0
 */

export interface MonthCount {
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  count: number;
}